Step 6 - We then create our react component, doing what ever we need. This example imports an free "password strength" package which we interface with.
We are listening for the change event on line 17 and 8 to update the actual form element with the value returned from the password package.

At the bottom we create a loop which maps the html tag in our blade file "password-meter" to the above component passing in a reference to the input field.

We should now have a fully working custom field built with react. If you're still running `npm run watch` this should have loaded for you.

We can confirm it's working in the next step.